server/requirements.txt
fastapi
uvicorn
numpy
usearch
openai
polars
tqdm
python-dotenv
